Add a .pgpass file for the opensrf user
This allows the opensrf user to psql into the database with `psql -h localhost -U Evergreen`,
without needing to enter the password.  This is handy for me when doing iterative development
with pgtap tests -- I can run the test quickly without needing to enter a password each time
I run it.