DEV container: Passing "yes" to npm ci commands
This dramatically decreases the time it takes for the container to boot. Angular has the annoying "yes/no" question when you first launch it. It times out after 10 minutes or something. But passing the "yes" into the command fixes the boot slowness.

These commands are fast when ran during docker build. But slow when ran during docker run. I think Docker is doing something in the environment during the build phase.