[dev image] Don't overwrite entityized fm_IDL.xml file when translation support is on
`cd build/i18n && make install` creates a very specific version of the fm_IDL.xml
file at Open-ILS/web/reports/fm_IDL.xml.  This version has a special "#include virtual"
doodad that allows it to pull in translations, and placeholders for the translated
strings to go into.  The code refers to this as an "entityized" file.

The project's `make install` process will install this into place at
/openils/var/web/reports/fm_IDL.xml, where it can be used to provide translated
names of fields and tables.

However, prior to this commit, both the evergreen_restart_services and
restart_post_boot playbooks subsequently copied
/openils/conf/fm_IDL.xml on top of the reports fm_IDL.xml, losing the ability
for Evergreen to provide the translated fields and tables.

This commit updates both playbooks to instead, if language support is turned on,
regenerate the entityized fm_IDL.xml to catch any hot new updates that the
developer has made locally and copy it into place in /openils/var/web/reports.