Copy IDL dtd file into place
To translate the IDL, the Makefile in build/i18n creates an
"entityized" IDL xml file with no strings whatsoever, along
with DTD files for each language that provide the necessary
strings.

Pull request #39 added the ability to regenerate the
entityized IDL and DTD files on each run of the
evergreen_restart_services or restart_post_boot playbooks.
It even copied the entityized IDL xml file into place!  But
it neglected to also copy the DTD files into place, which
had the potential to cause Internal Server Errors when
visiting /IDL2js.